+++ lib/src/in_app_webview/in_app_webview_settings.g.dart
@@ -1014,6 +1014,10 @@ class InAppWebViewSettings {
   ///- [InAppWebViewController.getScrollY]
   ///- [InAppWebViewController.scrollTo]
   ///- [InAppWebViewController.scrollBy]
+  ///- [InAppWebViewController.getContentHeight]
+  ///- [InAppWebViewController.getContentWidth]
+  ///- [InAppWebViewController.canScrollVertically]
+  ///- [InAppWebViewController.canScrollHorizontally]
   ///
   ///Settings affected:
   ///- [WebView.initialUserScripts]